Inserindo um no cliente com Post - Ajustado

+++ src/main/java/com/mmelero/cursomc/dto/ClienteNewDTO.java
@@ -20,7 +20,7 @@ public class ClienteNewDTO implements Serializable {
 	private String telefone2;
 	private String telefone3;
 
-	private Integer cidadeId;
+	private Long cidadeId;
 	
 	public ClienteNewDTO() {
 	}
@@ -121,11 +121,11 @@ public class ClienteNewDTO implements Serializable {
 		this.telefone3 = telefone3;
 	}
 
-	public Integer getCidadeId() {
+	public Long getCidadeId() {
 		return cidadeId;
 	}
 
-	public void setCidadeId(Integer cidadeId) {
+	public void setCidadeId(Long cidadeId) {
 		this.cidadeId = cidadeId;
 	}
 }
